Refactor file location
