+++ lib/aem/aemupload.js
@@ -31,7 +31,7 @@ const { Transfer } = require("../functions/transfer");
 const { executePipeline, Pipeline } = require("../generator/pipeline");
 const { RandomFileAccess } = require("../randomfileaccess");
 const EventEmitter = require("events");
-const UploadError = require("./upload-error");
+const UploadError = require("../block/upload-error");
 const { FilterFailedAssets } = require("../functions/filterfailedassets");
 const { IllegalArgumentError } = require("../error");
