+++ lib/block/blockupload.js
@@ -12,7 +12,6 @@
 
 'use strict';
 
-
 require("core-js/stable");
 
 const fileUrl = require("file-url");
@@ -30,7 +29,7 @@ const { executePipeline, Pipeline } = require("../generator/pipeline");
 const { RandomFileAccess } = require("../randomfileaccess");
 const EventEmitter = require("events");
 // TODO should the error file be in aem folder
-const UploadError = require("../aem/upload-error");
+const UploadError = require("./upload-error");
 const { FilterFailedAssets } = require("../functions/filterfailedassets");
 const { IllegalArgumentError } = require("../error");
 const { getFileStats } = require('../util');
