+++ lib/block/download-error.js
@@ -12,7 +12,7 @@
 
 'use strict';
 
-const errorCodes = require('../aem/error-codes');
+const errorCodes = require('./error-codes');
 
 /**
  * Concatenates to message values together if both are provided.
