+++ lib/block/error-codes.js
