+++ test/aem/upload-error.test.js
@@ -15,8 +15,8 @@
 'use strict';
 
 const assert = require("assert");
-const UploadError = require("../../lib/aem/upload-error");
-const errorCodes = require("../../lib/aem/error-codes");
+const UploadError = require("../../lib/block/upload-error");
+const errorCodes = require("../../lib/block/error-codes");
 
 describe('Upload Error', function() {
     function verifyCodes(httpStatus, expectedCode) {
